refactor: move common functions to helpers file
