+++ src/lib/quoteEngine.js
@@ -5,14 +5,20 @@ const { BigNumber, ethers } = require('ethers');
 const {
   NXM_PER_ALLOCATION_UNIT,
   ONE_YEAR,
-  PRICE_CHANGE_PER_DAY,
   SURGE_PRICE_RATIO,
   SURGE_THRESHOLD_DENOMINATOR,
   SURGE_THRESHOLD_RATIO,
   TARGET_PRICE_DENOMINATOR,
   SURGE_CHUNK_DIVISOR,
 } = require('./constants');
-const { calculateTrancheId, divCeil, bnMax, bnMin } = require('./helpers');
+const {
+  calculateFirstUsableTrancheIndex,
+  calculateBasePrice,
+  calculatePremiumPerYear,
+  calculateFixedPricePremiumPerYear,
+  divCeil,
+  bnMin,
+} = require('./helpers');
 const {
   selectAsset,
   selectAssetRate,
@@ -23,37 +29,6 @@ const {
 
 const { WeiPerEther, Zero, MaxUint256 } = ethers.constants;
 const { formatEther } = ethers.utils;
-
-const calculateBasePrice = (targetPrice, bumpedPrice, bumpedPriceUpdateTime, now) => {
-  const elapsed = now.sub(bumpedPriceUpdateTime);
-  const priceDrop = elapsed.mul(PRICE_CHANGE_PER_DAY).div(3600 * 24);
-  return bnMax(targetPrice, bumpedPrice.sub(priceDrop));
-};
-
-const calculateFixedPricePremiumPerYear = (coverAmount, price) => {
-  return coverAmount.mul(price).div(TARGET_PRICE_DENOMINATOR);
-};
-
-const calculatePremiumPerYear = (coverAmount, basePrice, initialCapacityUsed, totalCapacity) => {
-  const basePremium = coverAmount.mul(basePrice).div(TARGET_PRICE_DENOMINATOR);
-  const finalCapacityUsed = initialCapacityUsed.add(coverAmount);
-  const surgeStartPoint = totalCapacity.mul(SURGE_THRESHOLD_RATIO).div(SURGE_THRESHOLD_DENOMINATOR);
-
-  if (finalCapacityUsed.lte(surgeStartPoint)) {
-    return basePremium;
-  }
-
-  const amountOnSurgeSkip = initialCapacityUsed.sub(surgeStartPoint).gt(0)
-    ? initialCapacityUsed.sub(surgeStartPoint)
-    : Zero;
-
-  const amountOnSurge = finalCapacityUsed.sub(surgeStartPoint);
-  const totalSurgePremium = amountOnSurge.mul(amountOnSurge).mul(SURGE_PRICE_RATIO).div(totalCapacity).div(2);
-  const skipSurgePremium = amountOnSurgeSkip.mul(amountOnSurgeSkip).mul(SURGE_PRICE_RATIO).div(totalCapacity).div(2);
-  const surgePremium = totalSurgePremium.sub(skipSurgePremium);
-
-  return basePremium.add(surgePremium);
-};
 
 const calculatePoolPriceAndCapacity = (totalCapacity, basePrice, usedCapacity, useFixedPrice) => {
   const used = usedCapacity;
@@ -249,12 +224,7 @@ const quoteEngine = (store, productId, amount, period, coverAsset) => {
   const assetRates = store.getState().assetRates;
 
   const now = BigNumber.from(Date.now()).div(1000);
-  const gracePeriodExpiration = now.add(period).add(product.gracePeriod);
-
-  const firstActiveTrancheId = calculateTrancheId(now);
-  const firstUsableTrancheId = calculateTrancheId(gracePeriodExpiration);
-  const firstUsableTrancheIndex = firstUsableTrancheId - firstActiveTrancheId;
-
+  const firstUsableTrancheIndex = calculateFirstUsableTrancheIndex(now, product.gracePeriod, period);
   const coverAmountInNxm = amount.mul(WeiPerEther).div(assetRate);
 
   // rounding up to nearest allocation unit
